requirements.txt
numpy
pandas
google-cloud-bigquery
pandas-gbq
pyarrow